Ensure we use BUNDLE_WITHOUT
At the moment the BUNDLE_WITHOUT is being populated however the result of this was not actually being passed to the bundle install arguments. This looks to test if the BUNDLE_WITHOUT env variable is set and if it is pass it with the without flag to the bundle install command.