Fix definition of SCL related ENV variables and give ruby SCL bigger priority over nodejs
One important change is fixing the definition of the environment variables that are related
to SCL. It now more correspond with how it looks like when we just install and enable the Node.js
and Ruby collection.

This should also fix a bug that reported missing ../usr/local/bin path:

  https://bugzilla.redhat.com/show_bug.cgi?id=1909993

Ignoring PYTHONPATH is deliberate as it is set by the Node.js SCL only because of install
dependencies but if one would like to play with Python inside the Ruby container, it might
actually harm, so it's better to keep it unset and let use the system PYTHONPATH.

Swapping SCL names results in the latter having bigger priority, because PATH and other variables
are pre-ended with the new paths and list of SCLs are processed left to right.

Using rh-nodejsX rh-rubyY means that ruby SCL path will be before nodejs path in PATH variable
for example.