Check exit codes properly and remove test_base_scl_usage
Without set -e every command that can fail, including the test functions themselves, need to be explicitly checked for the return code.

The test test_base_scl_usage runs scl_usage function without args that obviously is not an expected scenario. It's not clear what the purpose of this test was, but it hardly does what it should.