Chown the directories added into container in test_from_dockerfile.
Since bundler v2.5.0, if the `frozen` option is not specified via CLI or config,
bundler will 'touch' the Gemfile.lock to update timestamps. This will fail
in test_from_dockerfile since the app-src is added into the container with
ownership of root:root.

Ideally, we expect the directory is writeable, so adjust ownership via the
Containerfile ADD --chown=<user>:<group> instruction.

Commit in bundler that introduced this behavior:
https://github.com/rubygems/rubygems/commit/b08f2f1208137c88f978500a281f5a74ee0aae02