deprecate ruby 3.1 for Fedora
No supported Fedora version provides Ruby 3.1. New Fedora image will
probably be added with adding Ruby 3.3.

The ruby-31 Fedora image was the only one we were building for
quay.io. Therefore whole build-and-push action is being removed with
this commit. We can re-add it when the ruby-33 for Fedora is ready.