Add example FIPS test application with HTTP.
This app has 4 GET endpoints that executes some OpenSSL capability.
2 of those test symmetric ciphers:
* '/symmetric/aes-256-cbc' -- succeeds under FIPS
* '/symmetric/des-ede-cbc' -- fails under FIPS
2 of those test digests:
* '/hash/sha256' -- succeeds under FIPS
* '/hash/md5' -- fails under FIPS

There is 5th endpoint for '/', that just executes `exit 0`
as a way to shut down the server.

These endpoints either return 200 if the case for them succeeded
or 5xx for FIPS related failures and 4xx for general failures not
accounted for. When a failure happens when it shouldn't, the app also
returns backtrace in the response body.

A few examples:
MD5 succeeds and FIPS is enabled, that's unexpected, returns 500
SHA256 fails in any case, 418 is returned because that shouldn't happen
with both FIPS disabled and enabled, something else went wrong.
MD5 fails with FIPS enabled, that's desired and expected, returns 200.

Since more information is passed within body on response,
`curl --fail-with-body` is recommended.

409 is chosen to differentiate 500 returned in cases we
might expect. It was chosen firstly as it is "user error", either the code is
wrong, or the setup is wrong.